[B] refactor: migrate MatchResult class
