+++ src/main/java/lotto/LottoConsoleGame.java
@@ -4,7 +4,7 @@ import lotto.lotto.Lotto;
 import lotto.lotto.LottoMachine;
 import lotto.money.Money;
 import lotto.number.WinningNumbers;
-import lotto.prize.MatchResult2;
+import lotto.prize.MatchResult;
 import lotto.view.InputView;
 import lotto.view.ResultView;
 
@@ -17,7 +17,7 @@ public class LottoConsoleGame {
         ResultView.printLottoTickets(lotto.getLottoTickets());
 
         WinningNumbers winningNumbers = InputView.inputWinningNumbers();
-        MatchResult2 matchResult = lotto.match(winningNumbers);
+        MatchResult matchResult = lotto.match(winningNumbers);
 
         ResultView.printMatchResult(matchResult);
     }
