+++ src/main/java/lotto/prize/MatchResult.java
@@ -10,16 +10,16 @@ import static lotto.money.Money.ZERO_VALUE;
 
 public class MatchResult {
     private final Money payment;
-    private final Map<LottoPrize, Long> matchResult;
+    private final Map<LottoPrizeTemp, Long> matchResult;
 
-    public MatchResult(Money payment, Map<LottoPrize, Long> matchResult) {
+    public MatchResult(Money payment, Map<LottoPrizeTemp, Long> matchResult) {
         validate(payment, matchResult);
 
         this.payment = payment;
         this.matchResult = matchResult;
     }
 
-    private void validate(Money payment, Map<LottoPrize, Long> matchResult) {
+    private void validate(Money payment, Map<LottoPrizeTemp, Long> matchResult) {
         if (Objects.isNull(payment)) {
             throw new IllegalArgumentException("Payment can't be null");
         }
@@ -29,24 +29,24 @@ public class MatchResult {
         }
     }
 
-    public static MatchResult of(Money payment, Map<LottoPrize, Long> matchResult) {
+    public static MatchResult of(Money payment, Map<LottoPrizeTemp, Long> matchResult) {
         return new MatchResult(payment, matchResult);
     }
 
     public double calculateEarningsRate() {
-        Money totalEarnings = Arrays.stream(LottoPrize.values())
+        Money totalEarnings = Arrays.stream(LottoPrizeTemp.values())
                 .map(this::calculateEarning)
                 .reduce(Money.ZERO, Money::addition);
 
         return totalEarnings.earningRate(payment);
     }
 
-    private Money calculateEarning(LottoPrize lottoPrize) {
+    private Money calculateEarning(LottoPrizeTemp lottoPrize) {
         return lottoPrize.getPrizeMoney()
                 .multiply(matchCount(lottoPrize));
     }
 
-    public int matchCount(LottoPrize lottoPrize) {
+    public int matchCount(LottoPrizeTemp lottoPrize) {
         if (!matchResult.containsKey(lottoPrize)) {
             return ZERO_VALUE;
         }
@@ -54,5 +54,4 @@ public class MatchResult {
         return matchResult.get(lottoPrize)
                 .intValue();
     }
-
 }
